requirements.txt
pyTelegramBotAPI
rembg
pillow